changed approach to model obj in cleanDb.ts

+++ server/src/seeds/cleanDb.ts
@@ -3,9 +3,15 @@ import db from '../config/connection.js';
 
 export default async (modelName: "Question", collectionName: string) => {
   try {
-    let modelExists = await models[modelName].db.db.listCollections({
+    const model :any|undefined = models[modelName];
+    
+    if (!model) {
+      throw new Error(`Model ${modelName} does not exist`);
+    }
+
+    let modelExists = await model.db.db.listCollections({
       name: collectionName
-    }).toArray()
+    }).toArray();
 
     if (modelExists.length) {
       await db.dropCollection(collectionName);
